+++ src/main/kotlin/us/frollo/frollosdk/database/SDKDatabase.kt
@@ -362,6 +362,7 @@ abstract class SDKDatabase : RoomDatabase() {
                 // 1) Update tables with new tracking_status values - budget, budget_period, goal, goal_period
                 // 2) Alter account table - add column features, product_available, cdr_p_product_id, cdr_p_product_name, cdr_p_product_details_page_url, cdr_p_key_information
                 // 3) Alter provider table - add column product_available
+                // 4) Alter transaction_model table - add column goal_id
 
                 database.execSQL(
                     "UPDATE budget  SET tracking_status = (CASE " +
@@ -418,6 +419,8 @@ abstract class SDKDatabase : RoomDatabase() {
                 database.execSQL("ALTER TABLE `account` ADD COLUMN `cdr_p_key_information` TEXT")
                 database.execSQL("ALTER TABLE `account` ADD COLUMN `products_available` INTEGER NOT NULL DEFAULT 0")
                 database.execSQL("ALTER TABLE `provider` ADD COLUMN `products_available` INTEGER NOT NULL DEFAULT 0")
+
+                database.execSQL("ALTER TABLE `transaction_model` ADD COLUMN `goal_id` INTEGER")
             }
         }
     }