+++ src/main/kotlin/us/frollo/frollosdk/model/api/aggregation/transactions/TransactionResponse.kt
@@ -41,5 +41,6 @@ internal data class TransactionResponse(
     @SerializedName("bill_id") var billId: Long?,
     @SerializedName("bill_payment_id") var billPaymentId: Long?,
     @SerializedName("user_tags") val userTags: List<String>?,
-    @SerializedName("external_id") val externalId: String?
+    @SerializedName("external_id") val externalId: String?,
+    @SerializedName("goal_id") val goalId: Long?
 )